Reorganize flask side
